requirements.txt
pydub
Pyrogram
tgcrypto
python_docx
Requests
